atlas/_requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
debugpy==1.8.1
Django==4.2.6
django-htmx==1.17.0
django-tailwind==3.6.0
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.2.0
Werkzeug==3.0.0
